+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './components/App';
-// import './index.css';
+import './index.css';
 
 ReactDOM.render(
   <App />,
